+++ packages/auto-complete-element/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - `aria-activedescendant` is empty when opening the list initially
 - `aria-multiselectable` was set on unfocusable element and hence screen readers wasn't announcing it
+- Input not emitting `change` event when modified
 
 ## [0.2.0] - 2022-06-01